Bootstrap custom variables

Use any Bootstrap variables to customize a shiny theme.

bs_vars(...)

bs_vars_global(
  body_bg = NULL,
  text_color = NULL,
  link_color = NULL,
  link_hover_color = NULL,
  line_height_base = NULL,
  grid_columns = NULL,
  grid_gutter_width = NULL,
  border_radius_base = NULL
)

bs_vars_color(
  brand_primary = NULL,
  brand_success = NULL,
  brand_info = NULL,
  brand_warning = NULL,
  brand_danger = NULL,
  gray_base = NULL,
  gray_darker = NULL,
  gray_dark = NULL,
  gray = NULL,
  gray_light = NULL,
  gray_lighter = NULL
)

bs_vars_navbar(
  height = NULL,
  margin_bottom = NULL,
  border_radius = NULL,
  padding_horizontal = NULL,
  padding_vertical = NULL,
  collapse_max_height = NULL,
  default_color = NULL,
  default_bg = NULL,
  default_border = NULL,
  default_link_color = NULL,
  default_link_active_color = NULL,
  default_link_active_bg = NULL,
  default_link_hover_color = NULL,
  default_link_hover_bg = NULL,
  inverse_color = NULL,
  inverse_bg = NULL,
  inverse_border = NULL,
  inverse_link_color = NULL,
  inverse_link_active_color = NULL,
  inverse_link_active_bg = NULL,
  inverse_link_hover_color = NULL,
  inverse_link_hover_bg = NULL
)

bs_vars_nav(
  link_padding = NULL,
  link_hover_bg = NULL,
  disabled_link_color = NULL,
  disabled_link_hover_color = NULL
)

bs_vars_font(
  family_sans_serif = NULL,
  size_base = NULL,
  size_large = NULL,
  size_small = NULL,
  size_h1 = NULL,
  size_h2 = NULL,
  size_h3 = NULL,
  size_h4 = NULL,
  size_h5 = NULL,
  size_h6 = NULL
)

bs_vars_wells(bg = NULL, border = NULL)

bs_vars_state(
  success_text = NULL,
  success_bg = NULL,
  success_border = NULL,
  info_text = NULL,
  info_bg = NULL,
  info_border = NULL,
  warning_text = NULL,
  warning_bg = NULL,
  warning_border = NULL,
  danger_text = NULL,
  danger_bg = NULL,
  danger_border = NULL
)

bs_vars_input(
  bg = NULL,
  color = NULL,
  border = NULL,
  border_radius = NULL,
  color_placeholder = NULL,
  group_addon_bg = NULL,
  border_focus = NULL,
  bg_disabled = NULL
)

bs_vars_alert(
  padding = NULL,
  border_radius = NULL,
  link_font_weight = NULL,
  success_text = NULL,
  success_bg = NULL,
  success_border = NULL,
  info_text = NULL,
  info_bg = NULL,
  info_border = NULL,
  warning_text = NULL,
  warning_bg = NULL,
  warning_border = NULL,
  danger_text = NULL,
  danger_bg = NULL,
  danger_border = NULL
)

bs_vars_progress(
  bg = NULL,
  bar_color = NULL,
  border_radius = NULL,
  bar_bg = NULL,
  bar_success_bg = NULL,
  bar_warning_bg = NULL,
  bar_danger_bg = NULL,
  bar_info_bg = NULL
)

bs_vars_panel(
  bg = NULL,
  body_padding = NULL,
  heading_padding = NULL,
  footer_padding = NULL,
  border_radius = NULL,
  inner_border = NULL,
  footer_bg = NULL,
  default_text = NULL,
  default_border = NULL,
  default_heading_bg = NULL,
  primary_text = NULL,
  primary_border = NULL,
  primary_heading_bg = NULL,
  success_text = NULL,
  success_border = NULL,
  success_heading_bg = NULL,
  info_text = NULL,
  info_border = NULL,
  info_heading_bg = NULL,
  warning_text = NULL,
  warning_border = NULL,
  warning_heading_bg = NULL,
  danger_text = NULL,
  danger_border = NULL,
  danger_heading_bg = NULL
)

bs_vars_modal(
  md = NULL,
  lg = NULL,
  sm = NULL,
  inner_padding = NULL,
  title_padding = NULL,
  title_line_height = NULL,
  content_bg = NULL,
  content_border_color = NULL,
  content_fallback_border_color = NULL,
  backdrop_bg = NULL,
  backdrop_opacity = NULL,
  header_border_color = NULL,
  footer_border_color = NULL
)

bs_vars_button(
  font_weight = NULL,
  default_color = NULL,
  default_bg = NULL,
  default_border = NULL,
  primary_color = NULL,
  primary_bg = NULL,
  primary_border = NULL,
  success_color = NULL,
  success_bg = NULL,
  success_border = NULL,
  info_color = NULL,
  info_bg = NULL,
  info_border = NULL,
  warning_color = NULL,
  warning_bg = NULL,
  warning_border = NULL,
  danger_color = NULL,
  danger_bg = NULL,
  danger_border = NULL,
  link_disabled_color = NULL,
  border_radius_base = NULL,
  border_radius_large = NULL,
  border_radius_small = NULL
)

bs_vars_dropdown(
  bg = NULL,
  border = NULL,
  fallback_border = NULL,
  divider_bg = NULL,
  link_color = NULL,
  link_hover_color = NULL,
  link_hover_bg = NULL,
  link_active_color = NULL,
  link_active_bg = NULL,
  link_disabled_color = NULL,
  header_color = NULL
)

bs_vars_pills(
  border_radius = NULL,
  active_link_hover_bg = NULL,
  active_link_hover_color = NULL
)

bs_vars_tabs(
  border_color = NULL,
  link_hover_border_color = NULL,
  active_link_hover_bg = NULL,
  active_link_hover_color = NULL,
  active_link_hover_border_color = NULL,
  justified_link_border_color = NULL,
  justified_active_link_border_color = NULL
)

bs_vars_badge(
  color = NULL,
  bg = NULL,
  link_hover_color = NULL,
  active_color = NULL,
  active_bg = NULL,
  font_weight = NULL,
  line_height = NULL,
  border_radius = NULL
)

bs_vars_component(
  padding_base_vertical = NULL,
  padding_base_horizontal = NULL,
  padding_large_vertical = NULL,
  padding_large_horizontal = NULL,
  padding_small_vertical = NULL,
  padding_small_horizontal = NULL,
  padding_xs_vertical = NULL,
  padding_xs_horizontal = NULL,
  line_height_large = NULL,
  line_height_small = NULL,
  border_radius_base = NULL,
  border_radius_large = NULL,
  border_radius_small = NULL,
  component_active_color = NULL,
  component_active_bg = NULL,
  caret_width_base = NULL,
  caret_width_large = NULL
)

bs_vars_table(
  cell_padding = NULL,
  condensed_cell_padding = NULL,
  bg = NULL,
  bg_accent = NULL,
  bg_hover = NULL,
  bg_active = NULL,
  border_color = NULL
)

Arguments

...

Variables to use, under the form body_bg = "#FFF" or "body-bg" = "#FFF".

body_bg

Background color for the body.

text_color

Global text color on body.

link_color

Dropdown link text color.

link_hover_color

Linked badge text color on hover.

line_height_base

Unit-less line-height for use in components like buttons.

grid_columns

Number of columns in the grid, e.g. in shiny::fluidRow(shiny::column(...)).

grid_gutter_width

Padding between columns. Gets divided in half for the left and right.

border_radius_base

Base border radius.

brand_primary

Primary color, default: #337ab7 .

brand_success

Success color, default: #5cb85c .

brand_info

Info color, default: #5bc0de .

brand_warning

Warning color, default: #f0ad4e .

brand_danger

Danger color, default: #d9534f .

gray_base

Base gray color.

gray_darker

Darker gray color.

gray_dark

Dark gray color.

gray

Gray color.

gray_light

Light gray color.

gray_lighter

Lighter gray color.

height

Height of the navbar, e.g. "50px" (the default in Bootstrap).

margin_bottom

Bottom margin of navbar.

border_radius

Border radius.

padding_horizontal

Horizontal padding.

padding_vertical

= Vertical padding.

collapse_max_height

Max height when collapsed.

default_color

Text color for default buttons.

default_bg

Background color for default buttons.

default_border

Border color for default buttons.

default_link_color

Link color.

default_link_active_color

Color for active link (selected tab).

default_link_active_bg

Background color for active link (selected tab).

default_link_hover_color

Color of links when hovered.

default_link_hover_bg

Background color of links when hovered.

inverse_color

Color of text for inverted navbar.

inverse_bg

Background color for inverted navbar.

inverse_border

Border color for inverted navbar.

inverse_link_color

Link color for inverted navbar.

inverse_link_active_color

Color for active link (selected tab) for inverted navbar.

inverse_link_active_bg

Background color for active link (selected tab) for inverted navbar.

inverse_link_hover_color

Color of links when hovered for inverted navbar.

inverse_link_hover_bg

Background color of links when hovered for inverted navbar.

link_padding

Padding for links (tabset's titles).

link_hover_bg

Hover background for dropdown links.

disabled_link_color

Disabled link color.

disabled_link_hover_color

Disabled link hover color.

family_sans_serif

Font family to use.

size_base

Size of base font, e.g. normal text, default in Bootstrap is "15px".

size_large

Size of large text.

size_small

Size of small text.

size_h1

Size of h1 tags.

size_h2

Size of h2 tags.

size_h3

Size of h3 tags.

size_h4

Size of h4 tags.

size_h5

Size of h5 tags.

size_h6

Size of h6 tags.

bg

Background color.

border

Dropdown menu border-color.

success_text

Text color for success status.

success_bg

Background color for success buttons.

success_border

Border color for success buttons.

info_text

Text color for info status.

info_bg

Background color for info buttons.

info_border

Border color for info buttons.

warning_text

Text color for warning status.

warning_bg

Background color for warning buttons.

warning_border

Border color for warning buttons.

danger_text

Text color for danger status.

danger_bg

Background color for danger buttons.

danger_border

Border color for danger buttons.

color

Text color.

color_placeholder

Text color of placeholder.

group_addon_bg

Background color of addons.

border_focus

Color of border when focused.

bg_disabled

Background color for disabled input.

padding

Padding for alerts.

link_font_weight

Font weight for links in alerts.

bar_color

Progress bar text color

bar_bg

Default progress bar color.

bar_success_bg

Success progress bar color.

bar_warning_bg

Warning progress bar color.

bar_danger_bg

Danger progress bar color.

bar_info_bg

Info progress bar color.

body_padding

Panel body padding.

heading_padding

Panel heading padding.

footer_padding

Panel footer padding.

inner_border

Border color for inner elements in panel.

footer_bg

Panel footer background color.

default_text

Default color for text.

default_heading_bg

Default background color for panel heading.

primary_text

Text color for primary status.

primary_border

Border color for primary buttons.

primary_heading_bg

Heading background color for primary status.

success_heading_bg

Heading background color for success status.

info_heading_bg

Heading background color for info status.

warning_heading_bg

Heading background color for warning status.

danger_heading_bg

Heading background color for danger status.

md

Size in pixel for medium modal, e.g. modalDialog(size = "m").

lg

Size in pixel for large modal, e.g. modalDialog(size = "l").

sm

Size in pixel for small modal, e.g. modalDialog(size = "s").

inner_padding

Padding applied to the modal body.

title_padding

Padding applied to the modal title.

title_line_height

Modal title line-height.

content_bg

Background color of modal content area.

content_border_color

Modal content border color.

content_fallback_border_color

Modal content border color (for IE8).

backdrop_bg

Modal backdrop background color.

backdrop_opacity

Modal backdrop opacity.

header_border_color

Modal header border color.

footer_border_color

Modal footer border color.

font_weight

Font weight, e.g. : "bold".

primary_color

Text color for primary buttons.

primary_bg

Background color for primary buttons.

success_color

Text color for success buttons.

info_color

Text color for info buttons.

warning_color

Text color for warning buttons.

danger_color

Text color for danger buttons.

link_disabled_color

Disabled dropdown menu item background color.

border_radius_large

Large border radius.

border_radius_small

Small border radius.

fallback_border

Dropdown menu border-color (for IE8).

divider_bg

Divider color for between dropdown items.

link_active_color

Active dropdown menu item text color.

link_active_bg

Active dropdown menu item background color.

header_color

Text color for headers within dropdown menus.

active_link_hover_bg

Active link hover background color.

active_link_hover_color

Active link hover color.

border_color

Border color.

link_hover_border_color

Link hover color.

active_link_hover_border_color

Active link hover border color.

justified_link_border_color

Justified link border color.

justified_active_link_border_color

Justified active link border color.

active_color

Badge text color in active nav link.

active_bg

Badge background color in active nav link.

line_height

Line height.

padding_base_vertical

Vertical base padding.

padding_base_horizontal

Horizontal base padding.

padding_large_vertical

Vertical large padding.

padding_large_horizontal

Horizontal large padding.

padding_small_vertical

Vertical small padding.

padding_small_horizontal

Horizontal small padding.

padding_xs_vertical

Vertical extra small padding.

padding_xs_horizontal

Horizontal extra small padding.

line_height_large

Line height for large elements.

line_height_small

Line height for small elements.

component_active_color

Color for active components.

component_active_bg

Background color for active components.

caret_width_base

Width for caret.

caret_width_large

Widget for large caret.

cell_padding

Cell padding.

condensed_cell_padding

Cell padding when using condensed table.

bg_accent

Background color used in striped table.

bg_hover

Background color used when hovering the table with the mouse.

bg_active

Background color when row is selected.

Value

a list that can be used in create_theme().

Note

For a full list of available variables, use search_vars_bs().

Examples

bs_vars(body_bg = "#FFF")
#> $`body-bg`
#> [1] "#FFF"
#> 
#> attr(,"class")
#> [1] "fresh_sass_vars" "bootstrap_vars"  "list"           

bs_vars("body-bg" = "#FFF")
#> $`body-bg`
#> [1] "#FFF"
#> 
#> attr(,"class")
#> [1] "fresh_sass_vars" "bootstrap_vars"  "list"