create_pretty {fresh}	R Documentation

Create a custom CSS file for pretty-checkbox

Description

This allow you to change colors of prettyCheckbox, prettyRadioButtons

Usage

create_pretty(
  output_file,
  default = NULL,
  primary = NULL,
  success = NULL,
  info = NULL,
  warning = NULL,
  danger = NULL
)

Arguments

output_file
Specifies path to output file for compiled CSS.

default
Default color.

primary
Primary color.

success
Success color.

info
Info color.

warning
Warning color.

danger
Danger color.

Value

If output_file = NULL, the function returns a string value of the compiled CSS. If the output path is specified, the compiled CSS is written to that file and invisible() is returned.

Examples

# Temporary file
tmp <- file.path(tempdir(), "my-pretty.css")

# Create the new theme
create_pretty(
  output_file = tmp,
  primary = "#FFFF00"
)

# Clean
unlink(tmp)

[Package fresh version 0.2.0 Index]