Create a custom CSS file for pretty-checkbox

This allow you to change colors of prettyCheckbox, prettyRadioButtons

create_pretty(
  output_file,
  default = NULL,
  primary = NULL,
  success = NULL,
  info = NULL,
  warning = NULL,
  danger = NULL
)

Arguments

output_file

Specifies path to output file for compiled CSS.

default

Default color.

primary

Primary color.

success

Success color.

info

Info color.

warning

Warning color.

danger

Danger color.

Value

If output_file = NULL, the function returns a string value of the compiled CSS. If the output path is specified, the compiled CSS is written to that file and invisible() is returned.

Examples

# Temporary file
tmp <- file.path(tempdir(), "my-pretty.css")

# Create the new theme
create_pretty(
  output_file = tmp,
  primary = "#FFFF00"
)
#> [1] "/tmp/Rtmp7N9yJH/my-pretty.css"

# Clean
unlink(tmp)