Create a custom Bootstrap theme

Allow to customize some CSS variables from Bootstrap themes to be included in Shiny applications.

create_theme(..., theme = c("default", "cerulean", "cosmo", "cyborg",
  "darkly", "flatly", "journal", "lumen", "paper", "readable", "sandstone",
  "simplex", "slate", "spacelab", "superhero", "united", "yeti"),
  output_file = NULL)

Arguments

...
Lists of CSS variables declared with bs_vars_* or adminlte_* functions.

theme
Base theme to use.

output_file
Specifies path to output file for compiled CSS.

Value

If output_file = NULL, the function returns a string value of the compiled CSS. If the output path is specified, the compiled CSS is written to that file and invisible() is returned.

Examples

# using a temporary file but use the path you want
tmp <- file.path(tempdir(), "custom-theme.css")

# Create the new theme
create_theme(
  theme = "default",
  bs_vars_color(
    brand_primary = "#75b8d1",
    brand_success = "#c9d175",
    brand_info = "#758bd1",
    brand_warning = "#d1ab75",
    brand_danger = "#d175b8"
  ),
  bs_vars_navbar(
    default_bg = "#75b8d1",
    default_color = "#FFFFFF",
    default_link_color = "#FFFFFF",
    default_link_active_color = "#FFFFFF"
  ),
  output_file = tmp
)

# Use the file created at the path provided
# in your Shiny app by moving it in the
# www/ folder, then use it in UI

library(shiny)
fluidPage(
  theme = "custom-theme.css"
)

#> <div class="container-fluid"></div>

# clean up
unlink(tmp)